requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dalex == 1.4.1
dice_ml == 0.9
ipython == 5.1.0
ipywidgets == 7.5.1
joblib == 1.1.1
matplotlib == 3.1.2
numpy == 1.17.2
pandas == 0.25.1
qii_tool == 0.1.3
raiwidgets == 0.23.0
responsibleai == 0.23.0
river == 0.8.0
scikit_learn == 0.24.2
scikit_multiflow == 0.5.3
scipy == 1.5.2
seaborn == 0.9.0
